+++ frontend/next.config.ts
@@ -4,6 +4,7 @@ const nextConfig: NextConfig = {
   output: "standalone",
   devIndicators: false, // enable/disable dev indicators 
   images: {
+    unoptimized: true,
     remotePatterns: [
       {
         protocol: "https",